Remove behavior from RangeSlider and SnapshotView control
