+++ controlsfx/src/main/java/impl/org/controlsfx/skin/RangeSliderSkin.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2013, 2016 ControlsFX
+ * Copyright (c) 2013, 2021, ControlsFX
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -26,41 +26,28 @@
  */
 package impl.org.controlsfx.skin;
 
-import static impl.org.controlsfx.behavior.RangeSliderBehavior.FocusedChild.HIGH_THUMB;
-import static impl.org.controlsfx.behavior.RangeSliderBehavior.FocusedChild.LOW_THUMB;
-import static impl.org.controlsfx.behavior.RangeSliderBehavior.FocusedChild.NONE;
-import static impl.org.controlsfx.behavior.RangeSliderBehavior.FocusedChild.RANGE_BAR;
-
-import java.util.List;
-
-import impl.org.controlsfx.behavior.RangeSliderBehavior;
-import impl.org.controlsfx.behavior.RangeSliderBehavior.FocusedChild;
 import javafx.beans.binding.ObjectBinding;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
 import javafx.event.EventHandler;
+import javafx.geometry.NodeOrientation;
 import javafx.geometry.Orientation;
 import javafx.geometry.Point2D;
 import javafx.geometry.Side;
 import javafx.scene.Cursor;
-import javafx.scene.Node;
 import javafx.scene.chart.NumberAxis;
+import javafx.scene.control.Skin;
 import javafx.scene.control.SkinBase;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.StackPane;
 import javafx.util.Callback;
-
 import org.controlsfx.control.RangeSlider;
+import org.controlsfx.tools.Utils;
 
-import com.sun.javafx.scene.traversal.Algorithm;
-import com.sun.javafx.scene.traversal.Direction;
-import com.sun.javafx.scene.traversal.ParentTraversalEngine;
-import com.sun.javafx.scene.traversal.TraversalContext;
+import static impl.org.controlsfx.skin.RangeSliderSkin.FocusedChild.*;
 
 public class RangeSliderSkin extends SkinBase<RangeSlider> {
-        
+
     /** Track if slider is vertical/horizontal and cause re layout */
     private NumberAxis tickLine = null;
     private double trackToTickGap = 2;
@@ -68,7 +55,7 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
     private boolean showTickMarks;
     private double thumbWidth;
     private double thumbHeight;
-    
+
     private Orientation orientation;
 
     private StackPane track;
@@ -80,19 +67,16 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
     private ThumbPane lowThumb;
     private ThumbPane highThumb;
     private StackPane rangeBar; // the bar between the two thumbs, can be dragged
-    
+
     // temp fields for mouse drag handling
     private double preDragPos;          // used as a temp value for low and high thumbs
     private Point2D preDragThumbPoint;  // in skin coordinates
-    
+
     private FocusedChild currentFocus = LOW_THUMB;
-    
-    private final RangeSliderBehavior rangeSliderBehavior;
 
     public RangeSliderSkin(final RangeSlider rangeSlider) {
         super(rangeSlider);
         orientation = getSkinnable().getOrientation();
-        rangeSliderBehavior = new RangeSliderBehavior(rangeSlider);
         initFirstThumb();
         initSecondThumb();
         initRangeBar();
@@ -142,163 +126,99 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
                 getSkinnable().requestLayout();
             }
         });
-        lowThumb.focusedProperty().addListener(new ChangeListener<Boolean>() {
-            @Override public void changed(ObservableValue<? extends Boolean> ov, Boolean t, Boolean hasFocus) {
-                if (hasFocus) {
-                    currentFocus = LOW_THUMB;
-                }
+        lowThumb.focusedProperty().addListener((ov, t, hasFocus) -> {
+            if (hasFocus) {
+                currentFocus = LOW_THUMB;
             }
         });
-        highThumb.focusedProperty().addListener(new ChangeListener<Boolean>() {
-            @Override public void changed(ObservableValue<? extends Boolean> ov, Boolean t, Boolean hasFocus) {
-                if (hasFocus) {
-                    currentFocus = HIGH_THUMB;
-                }
+        highThumb.focusedProperty().addListener((ov, t, hasFocus) -> {
+            if (hasFocus) {
+                currentFocus = HIGH_THUMB;
             }
         });
-        rangeBar.focusedProperty().addListener(new ChangeListener<Boolean>() {
-            @Override public void changed(ObservableValue<? extends Boolean> ov, Boolean t, Boolean hasFocus) {
-                if (hasFocus) {
-                    currentFocus = RANGE_BAR;
-                }
+        rangeBar.focusedProperty().addListener((ov, t, hasFocus) -> {
+            if (hasFocus) {
+                currentFocus = RANGE_BAR;
             }
         });
-        rangeSlider.focusedProperty().addListener(new ChangeListener<Boolean>() {
-            @Override public void changed(ObservableValue<? extends Boolean> ov, Boolean t, Boolean hasFocus) {
-                if (hasFocus) {
-                    lowThumb.setFocus(true);
-                } else {
-                    lowThumb.setFocus(false);
-                    highThumb.setFocus(false);
-                    currentFocus = NONE;
-                }
+        rangeSlider.focusedProperty().addListener((ov, t, hasFocus) -> {
+            if (hasFocus) {
+                lowThumb.setFocus(true);
+            } else {
+                lowThumb.setFocus(false);
+                highThumb.setFocus(false);
+                currentFocus = NONE;
             }
         });
 
-        EventHandler<KeyEvent> keyEventHandler = new EventHandler<KeyEvent>() {
-            /** The algorithm to be passed to
-             * {@link ParentTraversalEngine}{@code 's} constructor. */
-            private final Algorithm algorithm = new Algorithm() {
-                
-                @Override
-                public Node selectLast(final TraversalContext context) {
-                    final List<Node> focusTraversableNodes =
-                            context.getAllTargetNodes();
-                    
-                    return focusTraversableNodes.get(
-                            focusTraversableNodes.size() - 1);
+        EventHandler<KeyEvent> keyPressEventHandler = event -> {
+            if (KeyCode.TAB.equals(event.getCode())) {
+                if (lowThumb.isFocused()) {
+                    if (event.isShiftDown()) {
+                        focusPreviousChild();
+                    } else {
+                        lowThumb.setFocus(false);
+                        highThumb.setFocus(true);
+                    }
+                    event.consume();
+                } else if (highThumb.isFocused()) {
+                    if(event.isShiftDown()) {
+                        highThumb.setFocus(false);
+                        lowThumb.setFocus(true);
+                    } else {
+                        focusNextChild();
+                    }
+                    event.consume();
                 }
-                
-                
-                /**
-                 * Method is not used in current implementation. But in case of
-                 * using may return RangeSlider object itself instead of
-                 * {@link #lowThumb}, if RangeSlider is the first
-                 * focus traversable Node on the root pane.
-                 */
-                @Override
-                public Node selectFirst(final TraversalContext context) {
-                    return context.getAllTargetNodes().get(0);
+            } else if (KeyCode.LEFT.equals(event.getCode()) || KeyCode.KP_LEFT.equals(event.getCode())) {
+                if (getSkinnable().getOrientation() == Orientation.HORIZONTAL) {
+                    rtl(getSkinnable(), this::incrementValue, this::decrementValue);
                 }
-                
-                
-                /**
-                 * In case of passing {@link Direction#LEFT},
-                 * {@link Direction#PREVIOUS} or {@link Direction#UP} returns
-                 * Node <u>before</u> previous.
-                 * This method is called when pressing Shift + Tab on
-                 * the focused {@link #lowThumb}. The Node before 'lowThumb' is
-                 * RangeSlider object itself, so visually focus is lost and
-                 * focus traversing becomes distrupted.
-                 */
-                @Override
-                public Node select(
-                        final Node owner,
-                        final Direction dir,
-                        final TraversalContext context) {
-                    /* Determines focus moving direction to next
-                     * focus traversable node */
-                    final int direction;
-                    
-                    switch (dir)
-                    {
-                    case DOWN:
-                    case RIGHT:
-                    case NEXT:
-                    case NEXT_IN_LINE:
-                        direction = 1;
-                        
-                        break;
-                        
-                        
-                    case LEFT:
-                    case PREVIOUS:
-                    case UP:
-                        direction = -2;
-                        
-                        break;
-                        
-                        
-                    default:
-                        throw new EnumConstantNotPresentException(
-                                dir.getClass(), dir.name());
-                    }
-                    
-                    final List<Node> focusTraversableNodes =
-                            context.getAllTargetNodes();
-                    final int focusReceiverIndex =
-                            focusTraversableNodes.indexOf(owner) + direction;
-                    
-                    if (focusReceiverIndex < 0) {
-                        return focusTraversableNodes.get(
-                                focusTraversableNodes.size() - 1);
-                    }
-                    if (focusReceiverIndex == focusTraversableNodes.size()) {
-                        return focusTraversableNodes.get(0);
-                    }
-                    
-                    return focusTraversableNodes.get(focusReceiverIndex);
+            } else if (KeyCode.RIGHT.equals(event.getCode()) || KeyCode.KP_RIGHT.equals(event.getCode())) {
+                if (getSkinnable().getOrientation() == Orientation.HORIZONTAL) {
+                    rtl(getSkinnable(), this::decrementValue, this::incrementValue);
                 }
-            };
-            
-            @Override public void handle(KeyEvent event) {
-                if (KeyCode.TAB.equals(event.getCode())) {
-                    if (lowThumb.isFocused()) {
-                        if (event.isShiftDown()) {
-                            lowThumb.setFocus(false);
-                            new ParentTraversalEngine(
-                                    rangeSlider.getScene().getRoot(), algorithm)
-                            .select(lowThumb, Direction.PREVIOUS)
-                            .requestFocus();
-                        } else {
-                            lowThumb.setFocus(false);
-                            highThumb.setFocus(true);
-                        }
-                        event.consume();
-                    } else if (highThumb.isFocused()) {
-                        if(event.isShiftDown()) {
-                            highThumb.setFocus(false);
-                            lowThumb.setFocus(true);
-                        } else {
-                            highThumb.setFocus(false);
-                            new ParentTraversalEngine(
-                                    rangeSlider.getScene().getRoot(), algorithm)
-                            .select(highThumb, Direction.NEXT)
-                            .requestFocus();
-                        }
-                        event.consume();
-                    }
+            } else if (KeyCode.DOWN.equals(event.getCode()) || KeyCode.KP_DOWN.equals(event.getCode())) {
+                if (getSkinnable().getOrientation() == Orientation.VERTICAL) {
+                    decrementValue();
+                }
+            } else if (KeyCode.UP.equals(event.getCode()) || KeyCode.KP_UP.equals(event.getCode())) {
+                if (getSkinnable().getOrientation() == Orientation.VERTICAL) {
+                    incrementValue();
                 }
             }
+            event.consume();
         };
-        getSkinnable().addEventHandler(KeyEvent.KEY_PRESSED, keyEventHandler);  
-        // set up a callback on the behavior to indicate which thumb is currently 
-        // selected (via enum).
-        rangeSliderBehavior.setSelectedValue(new Callback<Void, FocusedChild>() {
-            @Override public FocusedChild call(Void v) {
-                return currentFocus;
+        EventHandler<KeyEvent> keyReleaseEventHandler = event -> {
+            if (KeyCode.HOME.equals(event.getCode())) {
+                home();
+            } else if (KeyCode.END.equals(event.getCode())) {
+                end();
             }
-        });
+            event.consume();
+        };
+        getSkinnable().addEventHandler(KeyEvent.KEY_PRESSED, keyPressEventHandler);
+        getSkinnable().addEventHandler(KeyEvent.KEY_RELEASED, keyReleaseEventHandler);
+        // set up a callback to indicate which thumb is currently selected (via enum).
+        setSelectedValue(v -> currentFocus);
+    }
+    
+    private void focusPreviousChild() {
+        final int index = getSkinnable().getParent().getChildrenUnmodifiable().indexOf(getSkinnable());
+        if (index != 0) {
+            getSkinnable().getParent().getChildrenUnmodifiable().get(index - 1).requestFocus();
+        } else {
+            getSkinnable().getParent().requestFocus();
+        }
+    }
+
+    private void focusNextChild() {
+        final int index = getSkinnable().getParent().getChildrenUnmodifiable().indexOf(getSkinnable());
+        if (index < getSkinnable().getParent().getChildrenUnmodifiable().size() - 1) {
+            getSkinnable().getParent().getChildrenUnmodifiable().get(index + 1).requestFocus();
+        } else {
+            getSkinnable().getParent().requestFocus();
+        }
     }
     
     private void initFirstThumb() {
@@ -306,126 +226,97 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
         lowThumb.getStyleClass().setAll("low-thumb"); //$NON-NLS-1$
         lowThumb.setFocusTraversable(true);
         track = new StackPane();
+        track.setFocusTraversable(false);
         track.getStyleClass().setAll("track"); //$NON-NLS-1$
 
         getChildren().clear();
         getChildren().addAll(track, lowThumb);
         setShowTickMarks(getSkinnable().isShowTickMarks(), getSkinnable().isShowTickLabels());
-        track.setOnMousePressed( new EventHandler<javafx.scene.input.MouseEvent>() {
-            @Override public void handle(javafx.scene.input.MouseEvent me) {
-                if (!lowThumb.isPressed() && !highThumb.isPressed()) {
-                    if (isHorizontal()) {
-                        rangeSliderBehavior.trackPress(me, (me.getX() / trackLength));
-                    } else {
-                        rangeSliderBehavior.trackPress(me, (me.getY() / trackLength));
-                    }
+        track.setOnMousePressed(me -> {
+            if (!lowThumb.isPressed() && !highThumb.isPressed()) {
+                if (isHorizontal()) {
+                    trackPress(me, (me.getX() / trackLength));
+                } else {
+                    trackPress(me, (me.getY() / trackLength));
                 }
             }
         });
 
-        track.setOnMouseReleased( new EventHandler<javafx.scene.input.MouseEvent>() {
-            @Override public void handle(javafx.scene.input.MouseEvent me) {
-                 //Nothing being done with the second param in sliderBehavior
-                //So, passing a dummy value
-                rangeSliderBehavior.trackRelease(me, 0.0f);
-            }
+        track.setOnMouseReleased(me -> {
+            // Nothing being done with the second param in sliderBehavior
+            // So, passing a dummy value
+            trackRelease(me, 0.0f);
         });
 
-        lowThumb.setOnMousePressed(new EventHandler<javafx.scene.input.MouseEvent>() {
-            @Override public void handle(javafx.scene.input.MouseEvent me) {
-                highThumb.setFocus(false);
-                lowThumb.setFocus(true);
-                rangeSliderBehavior.lowThumbPressed(me, 0.0f);
-                preDragThumbPoint = lowThumb.localToParent(me.getX(), me.getY());
-                preDragPos = (getSkinnable().getLowValue() - getSkinnable().getMin()) /
-                        (getMaxMinusMinNoZero());
-            }
+        lowThumb.setOnMousePressed(me -> {
+            highThumb.setFocus(false);
+            lowThumb.setFocus(true);
+            lowThumbPressed(me, 0.0f);
+            preDragThumbPoint = lowThumb.localToParent(me.getX(), me.getY());
+            preDragPos = (getSkinnable().getLowValue() - getSkinnable().getMin()) /
+                    (getMaxMinusMinNoZero());
         });
 
-        lowThumb.setOnMouseReleased(new EventHandler<javafx.scene.input.MouseEvent>() {
-            @Override public void handle(javafx.scene.input.MouseEvent me) {
-                rangeSliderBehavior.lowThumbReleased(me);
-            }
-        });
+        lowThumb.setOnMouseReleased(me -> lowThumbReleased(me));
 
-        lowThumb.setOnMouseDragged(new EventHandler<javafx.scene.input.MouseEvent>() {
-            @Override public void handle(javafx.scene.input.MouseEvent me) {
-                Point2D cur = lowThumb.localToParent(me.getX(), me.getY());
-                double dragPos = (isHorizontal())?
-                    cur.getX() - preDragThumbPoint.getX() : -(cur.getY() - preDragThumbPoint.getY());
-                rangeSliderBehavior.lowThumbDragged(me, preDragPos + dragPos / trackLength);
-            }
+        lowThumb.setOnMouseDragged(me -> {
+            Point2D cur = lowThumb.localToParent(me.getX(), me.getY());
+            double dragPos = (isHorizontal())?
+                cur.getX() - preDragThumbPoint.getX() : -(cur.getY() - preDragThumbPoint.getY());
+            lowThumbDragged(me, preDragPos + dragPos / trackLength);
         });
     }
 
     private void initSecondThumb() {
         highThumb = new ThumbPane();
         highThumb.getStyleClass().setAll("high-thumb"); //$NON-NLS-1$
-        highThumb.setFocusTraversable(true);
         if (!getChildren().contains(highThumb)) {
             getChildren().add(highThumb);
         }
+        highThumb.setOnMousePressed(e -> {
+            lowThumb.setFocus(false);
+            highThumb.setFocus(true);
+            highThumbPressed(e, 0.0D);
+            preDragThumbPoint = highThumb.localToParent(e.getX(), e.getY());
+            preDragPos = (getSkinnable().getHighValue() - getSkinnable().getMin()) / (getMaxMinusMinNoZero());
+        });
+        highThumb.setOnMouseReleased(e -> highThumbReleased(e));
+        highThumb.setOnMouseDragged(e -> {
+            boolean orientation = getSkinnable().getOrientation() == Orientation.HORIZONTAL;
+            double trackLength = orientation ? track.getWidth() : track.getHeight();
 
-        highThumb.setOnMousePressed(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                lowThumb.setFocus(false);
-                highThumb.setFocus(true);
-                rangeSliderBehavior.highThumbPressed(e, 0.0D);
-                preDragThumbPoint = highThumb.localToParent(e.getX(), e.getY());
-                preDragPos = (((RangeSlider) getSkinnable()).getHighValue() - ((RangeSlider) getSkinnable()).getMin()) / 
-                            (getMaxMinusMinNoZero());
-            }
-        }
-        );
-        highThumb.setOnMouseReleased(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                rangeSliderBehavior.highThumbReleased(e);
-            }
-        }
-        );
-        highThumb.setOnMouseDragged(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                boolean orientation = ((RangeSlider) getSkinnable()).getOrientation() == Orientation.HORIZONTAL;
-                double trackLength = orientation ? track.getWidth() : track.getHeight();
-
-                Point2D point2d = highThumb.localToParent(e.getX(), e.getY());
-                double d = ((RangeSlider) getSkinnable()).getOrientation() != Orientation.HORIZONTAL ? -(point2d.getY() - preDragThumbPoint.getY()) : point2d.getX() - preDragThumbPoint.getX();
-                rangeSliderBehavior.highThumbDragged(e, preDragPos + d / trackLength);
-            }
+            Point2D point2d = highThumb.localToParent(e.getX(), e.getY());
+            double d = getSkinnable().getOrientation() != Orientation.HORIZONTAL ? -(point2d.getY() - preDragThumbPoint.getY()) : point2d.getX() - preDragThumbPoint.getX();
+            highThumbDragged(e, preDragPos + d / trackLength);
         });
     }
-    
+
     private void initRangeBar() {
         rangeBar = new StackPane();
-        rangeBar.cursorProperty().bind(new ObjectBinding<Cursor>() {
-            { bind(rangeBar.hoverProperty()); }
+        rangeBar.setFocusTraversable(false);
+        rangeBar.cursorProperty().bind(new ObjectBinding<>() {
+            {
+                bind(rangeBar.hoverProperty());
+            }
 
-            @Override protected Cursor computeValue() {
+            @Override
+            protected Cursor computeValue() {
                 return rangeBar.isHover() ? Cursor.HAND : Cursor.DEFAULT;
             }
         });
         rangeBar.getStyleClass().setAll("range-bar"); //$NON-NLS-1$
-        
-        rangeBar.setOnMousePressed(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                rangeBar.requestFocus();
-                preDragPos = isHorizontal() ? e.getX() : -e.getY();
-            }
+
+        rangeBar.setOnMousePressed(e -> {
+            rangeBar.requestFocus();
+            preDragPos = isHorizontal() ? e.getX() : -e.getY();
         });
-        
-        rangeBar.setOnMouseDragged(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                double delta = (isHorizontal() ? e.getX() : -e.getY()) - preDragPos;
-                rangeSliderBehavior.moveRange(delta);
-            }
+
+        rangeBar.setOnMouseDragged(e -> {
+            double delta = (isHorizontal() ? e.getX() : -e.getY()) - preDragPos;
+            moveRange(delta);
         });
-        
-         rangeBar.setOnMouseReleased(new EventHandler<MouseEvent>() {
-            @Override public void handle(MouseEvent e) {
-                rangeSliderBehavior.confirmRange();
-            }
-        });
-        
+
+        rangeBar.setOnMouseReleased(e -> confirmRange());
         getChildren().add(rangeBar);
     }
 
@@ -443,6 +334,7 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
         if (showTickMarks) {
             if (tickLine == null) {
                 tickLine = new NumberAxis();
+                tickLine.setFocusTraversable(false);
                 tickLine.tickLabelFormatterProperty().bind(getSkinnable().labelFormatterProperty());
                 tickLine.setAnimated(false);
                 tickLine.setAutoRanging(false);
@@ -568,7 +460,7 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
             double tickLineWidth = (showTickMarks) ? tickLine.prefWidth(-1) : 0;
             double trackWidth = track.prefWidth(-1);
             double trackAreaWidth = Math.max(trackWidth,thumbWidth);
-            double totalWidthNeeded = trackAreaWidth  + ((showTickMarks) ? trackToTickGap+tickLineWidth : 0) ;
+            double totalWidthNeeded = trackAreaWidth  + ((showTickMarks) ? trackToTickGap + tickLineWidth : 0) ;
             double startX = x + ((w - totalWidthNeeded)/2); // center slider in available width horizontally
             trackLength = h - thumbHeight;
             trackStart = y + (thumbHeight/2);
@@ -598,7 +490,7 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
     }
     
     private double minTrackLength() {
-        return 2*lowThumb.prefWidth(-1);
+        return 2 * lowThumb.prefWidth(-1);
     }
     
     @Override protected double computeMinWidth(double height, double topInset, double rightInset, double bottomInset, double leftInset) {
@@ -664,9 +556,232 @@ public class RangeSliderSkin extends SkinBase<RangeSlider> {
         return orientation == null || orientation == Orientation.HORIZONTAL;
     }
 
+    /**************************************************************************
+     *                         State and Functions                            *
+     *************************************************************************/
+
+    private Callback<Void, FocusedChild> selectedValue;
+    public void setSelectedValue(Callback<Void, FocusedChild> c) {
+        selectedValue = c;
+    }
+
+    // when high thumb is released, highValueChanging is set to false.
+    private void highThumbReleased(MouseEvent e) {
+        RangeSlider slider = getSkinnable();
+        slider.setHighValueChanging(false);
+        if (slider.isSnapToTicks())
+            slider.setHighValue(snapValueToTicks(slider.getHighValue()));
+    }
+
+    private void highThumbPressed(MouseEvent e, double position) {
+        RangeSlider slider = getSkinnable();
+        if (!slider.isFocused())
+            slider.requestFocus();
+        slider.setHighValueChanging(true);
+    }
+
+    private void highThumbDragged(MouseEvent e, double position) {
+        RangeSlider slider = getSkinnable();
+        slider.setHighValue(Utils.clamp(slider.getMin(), position * (slider.getMax() - slider.getMin()) + slider.getMin(), slider.getMax()));
+    }
+
+    private void moveRange(double position) {
+        RangeSlider slider = getSkinnable();
+        final double min = slider.getMin();
+        final double max = slider.getMax();
+        final double lowValue = slider.getLowValue();
+        final double newLowValue = Utils.clamp(min, lowValue + position *(max-min) /
+                (slider.getOrientation() == Orientation.HORIZONTAL? slider.getWidth(): slider.getHeight()), max);
+        final double highValue = slider.getHighValue();
+        final double newHighValue = Utils.clamp(min, highValue + position*(max-min) /
+                (slider.getOrientation() == Orientation.HORIZONTAL? slider.getWidth(): slider.getHeight()), max);
+
+        if (newLowValue <= min || newHighValue >= max) return;
+        slider.setLowValueChanging(true);
+        slider.setHighValueChanging(true);
+        slider.setLowValue(newLowValue);
+        slider.setHighValue(newHighValue);
+    }
+
+    private void confirmRange() {
+        RangeSlider slider = getSkinnable();
+
+        slider.setLowValueChanging(false);
+        if (slider.isSnapToTicks()) {
+            slider.setLowValue(snapValueToTicks(slider.getLowValue()));
+        }
+        slider.setHighValueChanging(false);
+        if (slider.isSnapToTicks()) {
+            slider.setHighValue(snapValueToTicks(slider.getHighValue()));
+        }
+    }
+
+    /**
+     * Invoked by the RangeSlider {@link Skin} implementation whenever a mouse press
+     * occurs on the "track" of the slider. This will cause the thumb to be
+     * moved by some amount.
+     *
+     * @param position The mouse position on track with 0.0 being beginning of
+     *        track and 1.0 being the end
+     */
+    private void trackPress(MouseEvent e, double position) {
+        // determine the percentage of the way between min and max
+        // represented by this mouse event
+        final RangeSlider rangeSlider = getSkinnable();
+        // If not already focused, request focus
+        if (!rangeSlider.isFocused()) {
+            rangeSlider.requestFocus();
+        }
+        if (selectedValue != null) {
+            double newPosition;
+            if (rangeSlider.getOrientation().equals(Orientation.HORIZONTAL)) {
+                newPosition = position * (rangeSlider.getMax() - rangeSlider.getMin()) + rangeSlider.getMin();
+            } else {
+                newPosition = (1 - position) * (rangeSlider.getMax() - rangeSlider.getMin()) + rangeSlider.getMin();
+            }
+
+            /**
+             * If the position is inferior to the current LowValue, this means
+             * the user clicked on the track to move the low thumb. If not, then
+             * it means the user wanted to move the high thumb.
+             */
+            if (newPosition < rangeSlider.getLowValue()) {
+                rangeSlider.adjustLowValue(newPosition);
+            } else {
+                rangeSlider.adjustHighValue(newPosition);
+            }
+        }
+    }
+
+    public void trackRelease(MouseEvent e, double position) {
+    }
+
+    /**
+     * @param position The mouse position on track with 0.0 being beginning of
+     *       track and 1.0 being the end
+     */
+    public void lowThumbPressed(MouseEvent e, double position) {
+        // If not already focused, request focus
+        final RangeSlider rangeSlider = getSkinnable();
+        if (!rangeSlider.isFocused())  rangeSlider.requestFocus();
+        rangeSlider.setLowValueChanging(true);
+    }
+
+    /**
+     * @param position The mouse position on track with 0.0 being beginning of
+     *        track and 1.0 being the end
+     */
+    public void lowThumbDragged(MouseEvent e, double position) {
+        final RangeSlider rangeSlider = getSkinnable();
+        double newValue = Utils.clamp(rangeSlider.getMin(),
+                (position * (rangeSlider.getMax() - rangeSlider.getMin())) + rangeSlider.getMin(),
+                rangeSlider.getMax());
+        rangeSlider.setLowValue(newValue);
+    }
+
+    /**
+     * When lowThumb is released lowValueChanging should be set to false.
+     */
+    public void lowThumbReleased(MouseEvent e) {
+        final RangeSlider rangeSlider = getSkinnable();
+        rangeSlider.setLowValueChanging(false);
+        // RT-15207 When snapToTicks is true, slider value calculated in drag
+        // is then snapped to the nearest tick on mouse release.
+        if (rangeSlider.isSnapToTicks()) {
+            rangeSlider.setLowValue(snapValueToTicks(rangeSlider.getLowValue()));
+        }
+    }
+
+    void home() {
+        RangeSlider slider = (RangeSlider) getNode();
+        slider.adjustHighValue(slider.getMin());
+    }
+
+    private void decrementValue() {
+        RangeSlider slider = getSkinnable();
+        if (selectedValue != null) {
+            if (selectedValue.call(null) == FocusedChild.HIGH_THUMB) {
+                if (slider.isSnapToTicks())
+                    slider.adjustHighValue(slider.getHighValue() - computeIncrement());
+                else
+                    slider.decrementHighValue();
+            } else {
+                if (slider.isSnapToTicks())
+                    slider.adjustLowValue(slider.getLowValue() - computeIncrement());
+                else
+                    slider.decrementLowValue();
+            }
+        }
+    }
+
+    private void end() {
+        RangeSlider slider = (RangeSlider) getNode();
+        slider.adjustHighValue(slider.getMax());
+    }
+
+    private void incrementValue() {
+        RangeSlider slider = getSkinnable();
+        if (selectedValue != null) {
+            if (selectedValue.call(null) == FocusedChild.HIGH_THUMB) {
+                if (slider.isSnapToTicks())
+                    slider.adjustHighValue(slider.getHighValue() + computeIncrement());
+                else
+                    slider.incrementHighValue();
+            } else {
+                if (slider.isSnapToTicks())
+                    slider.adjustLowValue(slider.getLowValue() + computeIncrement());
+                else
+                    slider.incrementLowValue();
+            }
+        }
+    }
+
+    private double computeIncrement() {
+        RangeSlider rangeSlider = getSkinnable();
+        double d;
+        if (rangeSlider.getMinorTickCount() != 0)
+            d = rangeSlider.getMajorTickUnit() / (double) (Math.max(rangeSlider.getMinorTickCount(), 0) + 1);
+        else
+            d = rangeSlider.getMajorTickUnit();
+        if (rangeSlider.getBlockIncrement() > 0.0D && rangeSlider.getBlockIncrement() < d)
+            return d;
+        else
+            return rangeSlider.getBlockIncrement();
+    }
+
+    private void rtl(RangeSlider node, Runnable rtlMethod, Runnable nonRtlMethod) {
+        if (node.getEffectiveNodeOrientation() == NodeOrientation.RIGHT_TO_LEFT) {
+            rtlMethod.run();
+        } else {
+            nonRtlMethod.run();
+        }
+    }
+
+    private double snapValueToTicks(double d) {
+        RangeSlider rangeSlider = getSkinnable();
+        double d1 = d;
+        double d2;
+        if (rangeSlider.getMinorTickCount() != 0)
+            d2 = rangeSlider.getMajorTickUnit() / (double) (Math.max(rangeSlider.getMinorTickCount(), 0) + 1);
+        else
+            d2 = rangeSlider.getMajorTickUnit();
+        int i = (int) ((d1 - rangeSlider.getMin()) / d2);
+        double d3 = (double) i * d2 + rangeSlider.getMin();
+        double d4 = (double) (i + 1) * d2 + rangeSlider.getMin();
+        d1 = Utils.nearest(d3, d1, d4);
+        return Utils.clamp(rangeSlider.getMin(), d1, rangeSlider.getMax());
+    }
+
     private static class ThumbPane extends StackPane {
         public void setFocus(boolean value) {
             setFocused(value);
         }
+    }
+
+    public enum FocusedChild {
+        LOW_THUMB,
+        HIGH_THUMB,
+        RANGE_BAR,
+        NONE
     }
 }
